+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release "Feige", Unreleased
 
 ## Features
+- Allow users to remove trust bananas on their own profile #592 !1920 @alex.simm
 
 ## Changes
 - Profile storelist now shows store cooperation status !1828 @ChrisOelmueller
@@ -37,7 +38,6 @@
 - New menu entry for changing the language of the website #1015 !1877 !1880 @alex.simm
 - Show icon and tooltip for working groups with extra functions !1846 @alex.simm
 - Added a button that allows the creator of a poll to delete it within the first hour #975 !1906 !1912 @alex.simm
-- Allow users to remove trust bananas on their own profile #592 @alex.simm
 
 ## Changes
 - New look for mailbox folder list (rewrite in Vue) !1788 @ChrisOelmueller